backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
simple-salesforce==1.12.5
python-dotenv==1.0.0
groq>=0.13.0
pandas==2.0.0
numpy==1.24.3
openpyxl==3.1.5
pydantic==2.5.0
requests==2.31.0
httpx==0.27.0
firebase-admin==6.5.0
PyJWT==2.8.0
